+++ src/app/dashboard/[orgID]/settings/page.tsx
@@ -1,8 +1,7 @@
 import ContentWrapper from "@/components/dashboard/ContentWrapper";
 import Main from "@/components/dashboard/MainContainer";
-import { General, Header, Navigation } from "./Components";
-
-
+import Header from "@/components/dashboard/Header/Header";
+import { General, Navigation } from "./Components";
 
 export default function Page() {
   return (
