+++ src/app/dashboard/[orgID]/tasks/Components/TaskDetails/Details.tsx
@@ -1,0 +1,9 @@
+import styles from './TasksGrid.module.scss';
+
+export default function Details() {
+    return (
+        <div>
+            Task Details
+        </div>
+    );
+}
