+++ src/app/dashboard/[orgID]/tasks/Components/index.ts
@@ -1,0 +1,4 @@
+export { default as TaskDetails } from './TaskDetails/Details';
+export { default as TaskGrid } from './TasksGrid/TasksGrid';
+export { default as ProjectList } from './Projects/ProjectList';
+export { default as Content } from './Content/Content';
