+++ src/app/dashboard/[orgID]/tasks/[taskID]/page.tsx
@@ -1,5 +1,20 @@
-export default function Task() {
+"use client";
+
+import ContentWrapper from "@/components/dashboard/ContentWrapper";
+import Main from "@/components/dashboard/MainContainer";
+import Header from "@/components/dashboard/Header/Header";
+import { ProjectList, Content } from "../Components";
+
+export default function Page() {
   return (
-    <h1>Task</h1>
+    <div className="dd-content">
+      <ProjectList />
+      <Header />
+      <Main>
+        <ContentWrapper>
+          <Content />
+        </ContentWrapper>
+      </Main>
+    </div>
   );
 }
